pyproject.toml
[project]
name = "space-eval-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "absl-py>=2.4.0",
    "datasets>=4.0.0",
    "evaluate>=0.4.6",
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.8",
    "nltk>=3.9.2",
    "pyyaml>=6.0.2",
    "rouge-score>=0.1.2",
    "scikit-learn>=1.8.0",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "vllm>=0.15.0",
]
